+++ src/GafferScene/ShaderPlug.cpp
@@ -40,8 +40,12 @@
 
 #include "Gaffer/BoxIO.h"
 #include "Gaffer/Context.h"
+#include "Gaffer/ContextProcessor.h"
 #include "Gaffer/Dot.h"
+#include "Gaffer/Loop.h"
+#include "Gaffer/PlugAlgo.h"
 #include "Gaffer/ScriptNode.h"
+#include "Gaffer/Spreadsheet.h"
 #include "Gaffer/SubGraph.h"
 #include "Gaffer/Switch.h"
 
@@ -102,6 +106,9 @@ bool isParameterType( const Plug *plug )
 	}
 }
 
+const IECore::InternedString hasProxyNodesIdentifier( "__hasProxyNodes" );
+const IECore::InternedString g_out( "out" );
+
 } // namespace
 
 //////////////////////////////////////////////////////////////////////////
@@ -157,8 +164,18 @@ bool ShaderPlug::acceptsInput( const Gaffer::Plug *input ) const
 		return true;
 	}
 
-	// And we support switches by traversing across them
-	// ourselves when necessary, in `shaderOutPlug()`.
+	// We also allow a bunch of general-purpose nodes below, but there are
+	// certain plug types we know a shader can never be connected to. Reject
+	// those, otherwise stupid things happen, like the ShaderView trying to
+	// display scenes and images.
+
+	if( !isParameterType( sourcePlug ) )
+	{
+		return false;
+	}
+
+	// Allow nodes we handle using `PlugAlgo::contextSensitiveSource()` in
+	// `sourceShader()`.
 
 	const Node *sourceNode = sourcePlug->node();
 	if( auto switchNode = runTimeCast<const Switch>( sourceNode ) )
@@ -177,17 +194,23 @@ bool ShaderPlug::acceptsInput( const Gaffer::Plug *input ) const
 					return false;
 				}
 			}
-			// For switches without any inputs, we have to assume that
-			// the user might connect a shader in the future. But there
-			// are certain plug types we know a shader can never be connected
-			// to. Reject those, otherwise stupid things happen, like the
-			// ShaderView trying to display scenes and images.
-			if( !isParameterType( sourcePlug ) )
-			{
-				return false;
-			}
 			return true;
 		}
+	}
+	else if(
+		runTimeCast<const ContextProcessor>( sourceNode ) ||
+		runTimeCast<const Loop>( sourceNode ) ||
+		runTimeCast<const Spreadsheet>( sourceNode )
+	)
+	{
+		/// \todo Ideally we'd also check that `sourceNode.in` doesn't have an
+		/// input that we can't accept ourselves. But there's nothing stopping
+		/// someone connecting one later even if there isn't one now, so we just
+		/// have to reject them later in `sourceShader()` instead. The only
+		/// reason we can do it for Switch (above) is that Switch overrides
+		/// `acceptsInput()` in a way that would reject the later connection.
+		auto sourceNodeOut = sourceNode->getChild<Plug>( g_out );
+		return sourceNodeOut && ( sourcePlug == sourceNodeOut || sourceNodeOut->isAncestorOf( sourcePlug ) );
 	}
 
 	// We must accept intermediate connections from plugs on utility nodes on the
@@ -201,31 +224,31 @@ bool ShaderPlug::acceptsInput( const Gaffer::Plug *input ) const
 		runTimeCast<const BoxIO>( sourceNode )
 	)
 	{
-		if( isParameterType( sourcePlug ) )
-		{
-			return true;
-		}
+		return true;
 	}
 
 	return false;
 }
 
+struct ShaderPlug::ShaderContext
+{
+
+	ConstContextPtr context;
+	std::optional<Context::Scope> scope;
+	std::string outputParameter;
+	std::optional<Context::EditableScope> editableScope;
+
+};
+
 IECore::MurmurHash ShaderPlug::attributesHash() const
 {
 	IECore::MurmurHash h;
-	if( const Gaffer::Plug *p = shaderOutPlug() )
+	ShaderContext context;
+	const Gaffer::Plug *plug = shaderOutPlug( context );
+	if( plug )
 	{
-		if( auto s = runTimeCast<const GafferScene::Shader>( p->node() ) )
-		{
-			Context::EditableScope scope( Context::current() );
-			std::string outputParameter;
-			if( p != s->outPlug() )
-			{
-				outputParameter = p->relativeName( s->outPlug() );
-				scope.set( Shader::g_outputParameterContextName, &outputParameter );
-			}
-			h = s->outAttributesPlug()->hash();
-		}
+		auto shader = static_cast<const Shader*>( plug->node() );
+		h = shader->outAttributesPlug()->hash();
 	}
 
 	return h;
@@ -233,56 +256,50 @@ IECore::MurmurHash ShaderPlug::attributesHash() const
 
 IECore::ConstCompoundObjectPtr ShaderPlug::attributes() const
 {
-	static IECore::InternedString hasProxyNodesIdentifier( "__hasProxyNodes" );
-	if( const Gaffer::Plug *p = shaderOutPlug() )
+	ShaderContext context;
+	const Gaffer::Plug *plug = shaderOutPlug( context );
+	if( plug )
 	{
-		if( auto s = runTimeCast<const GafferScene::Shader>( p->node() ) )
+		auto shader = static_cast<const Shader*>( plug->node() );
+		IECore::ConstCompoundObjectPtr result = shader->outAttributesPlug()->getValue();
+
+		// Check for outputs from ShaderTweakProxy, which should only be used with ShaderTweaks nodes
+		for( const auto &i : result->members() )
 		{
-			Context::EditableScope scope( Context::current() );
-			std::string outputParameter;
-			if( p != s->outPlug() )
+			if( const IECoreScene::ShaderNetwork *shaderNetwork = IECore::runTimeCast< const IECoreScene::ShaderNetwork >( i.second.get() ) )
 			{
-				outputParameter = p->relativeName( s->outPlug() );
-				scope.set( Shader::g_outputParameterContextName, &outputParameter );
-			}
-
-			IECore::ConstCompoundObjectPtr result = s->outAttributesPlug()->getValue();
-
-			// Check for outputs from ShaderTweakProxy, which should only be used with ShaderTweaks nodes
-			for( const auto &i : result->members() )
-			{
-				if( const IECoreScene::ShaderNetwork *shaderNetwork = IECore::runTimeCast< const IECoreScene::ShaderNetwork >( i.second.get() ) )
+				if( const BoolData *hasProxyNodes = shaderNetwork->blindData()->member<IECore::BoolData>( hasProxyNodesIdentifier ) )
 				{
-					if( const BoolData *hasProxyNodes = shaderNetwork->blindData()->member<IECore::BoolData>( hasProxyNodesIdentifier ) )
+					if( hasProxyNodes->readable() )
 					{
-						if( hasProxyNodes->readable() )
-						{
-							throw IECore::Exception(
-								"ShaderTweakProxy only works with ShaderTweaks - it doesn't make sense to connect one here"
-							);
+						throw IECore::Exception(
+							"ShaderTweakProxy only works with ShaderTweaks - it doesn't make sense to connect one here"
+						);
 
-						}
 					}
 				}
 			}
-			return result;
 		}
+
+		return result;
 	}
+
 	return new CompoundObject;
 }
 
 const Gaffer::ValuePlug *ShaderPlug::parameterSource( const IECoreScene::ShaderNetwork::Parameter &parameter) const
 {
-	if( const Gaffer::Plug *p = shaderOutPlug() )
+	ShaderContext context;
+	const Gaffer::Plug *plug = shaderOutPlug( context );
+
+	if( plug )
 	{
-		if( auto s = runTimeCast<const GafferScene::Shader>( p->node() ) )
+		auto shader = static_cast<const Shader*>( plug->node() );
+		if( parameter.shader.string().empty() )
 		{
-			if( parameter.shader.string().empty() )
-			{
-				return s->parametersPlug()->descendant<ValuePlug>( parameter.name );
-			}
-			return s->parameterSource( p, parameter );
+			return shader->parametersPlug()->descendant<ValuePlug>( parameter.name );
 		}
+		return shader->parameterSource( plug, parameter );
 	}
 	return nullptr;
 }
@@ -292,28 +309,40 @@ Gaffer::ValuePlug *ShaderPlug::parameterSource( const IECoreScene::ShaderNetwork
 	return const_cast<ValuePlug *>( const_cast<const ShaderPlug *>( this )->parameterSource( parameter ) );
 }
 
-const Gaffer::Plug *ShaderPlug::shaderOutPlug() const
+const Gaffer::Plug *ShaderPlug::shaderOutPlug( ShaderContext &shaderContext ) const
 {
-	const Plug *source = this->source<Gaffer::Plug>();
+	auto [source, context] = PlugAlgo::contextSensitiveSource( this );
 	if( source == this )
 	{
 		// No input
 		return nullptr;
 	}
 
-	if( auto switchNode = runTimeCast<const Switch>( source->node() ) )
+	auto shader = runTimeCast<const GafferScene::Shader>( source->node() );
+	if( !shader )
 	{
-		// Special case for switches with context-varying index values.
-		// Query the active input for this context, and manually traverse
-		// out the other side.
-		/// \todo Perhaps we should support ContextProcessors in the same way?
-		/// We have a similar pattern now in ShaderPlug, Shader::NetworkBuilder
-		/// and Dispatcher. Perhaps the logic should be consolidated into a
-		/// `PlugAlgo::computedSource()` utility of some sort?
-		if( const Plug *activeInPlug = switchNode->activeInPlug( source ) )
-		{
-			source = activeInPlug->source();
-		}
+		return nullptr;
+	}
+
+	const Plug *shaderOutPlug = shader->outPlug();
+	if(
+		!shaderOutPlug ||
+		( source != shaderOutPlug && !shaderOutPlug->isAncestorOf( source ) )
+	)
+	{
+		return nullptr;
+	}
+
+	if( source != shaderOutPlug )
+	{
+		shaderContext.editableScope.emplace( context.get() );
+		shaderContext.outputParameter = source->relativeName( shaderOutPlug );
+		shaderContext.editableScope->set( Shader::g_outputParameterContextName, &shaderContext.outputParameter );
+	}
+	else if( context != Context::current() )
+	{
+		shaderContext.context = context;
+		shaderContext.scope.emplace( context.get() );
 	}
 
 	return source;