+++ Changes.md
@@ -9,7 +9,8 @@ Features
 Improvements
 ------------
 
-- Shader : Added support for ContextProcessor, Loop and Spreadsheet nodes to be used inline between shader nodes.
+- Shader, ShaderPlug : Added support for ContextProcessor, Loop and Spreadsheet nodes to be used inline between shader nodes and as the terminal node connected to
+  ShaderAssignment and other shader-consuming nodes.
 - VisualiserTool : Changed `dataName` input widget for choosing the primitive variable to visualise to a list of available variable names for the current selection.
 - Tweaks nodes : Moved list of tweaks to a collapsible "Tweaks" section in the NodeEditor.
 - Viewer :